Explicitly cast data tests to precise types
Previously the real number tests for being implicitly casted to SQL
exact numeric types, which can have precision exceeding R numeric types.
For this reason they were being converted to strings, which was causing
the tests to fail.

This change explicit defines them to be real types, which can be
exactly converted to R numeric types.